improved setting up the calender, got rid of duplicate code in all of the fragments
