+++ app/src/main/java/io/punchtime/punchtime/ui/fragments/CalendarSetup.java
@@ -1,0 +1,138 @@
+package io.punchtime.punchtime.ui.fragments;
+
+
+import android.graphics.RectF;
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.support.v4.content.ContextCompat;
+import android.view.View;
+
+import com.alamkanak.weekview.MonthLoader;
+import com.alamkanak.weekview.WeekView;
+import com.alamkanak.weekview.WeekViewEvent;
+import com.firebase.client.DataSnapshot;
+import com.firebase.client.Firebase;
+import com.firebase.client.FirebaseError;
+import com.firebase.client.Query;
+import com.firebase.client.ValueEventListener;
+
+import java.util.ArrayList;
+import java.util.Calendar;
+import java.util.List;
+
+import io.punchtime.punchtime.R;
+import io.punchtime.punchtime.data.Pulse;
+import io.punchtime.punchtime.ui.activities.MainActivity;
+
+/**
+ * Created by arnaud on 05/05/16.
+ */
+public class CalendarSetup extends Fragment implements WeekView.EventClickListener, MonthLoader.MonthChangeListener,
+        WeekView.EventLongPressListener, WeekView.EmptyViewLongPressListener {
+    private WeekView mWeekView;
+    private Firebase mRef;
+    private MainActivity activity;
+    List<Pulse> pulseList;
+
+
+    // triggered soon after onCreateView
+    // Any view setup should occur here.
+    @Override
+    public void onViewCreated(View view, Bundle savedInstanceState) {
+        // Setup any handles to view objects
+        Calendar cal = Calendar.getInstance();
+        mWeekView.goToHour(cal.get(Calendar.HOUR_OF_DAY) - 1);
+    }
+
+    // gets the note for a given pulse
+    protected String getEventTitle(Pulse pulse) {
+        return pulse.getAddressStreet() + "\n" + pulse.getAddressCityCountry();
+    }
+
+    // Gets pulses ASYNC from Firebase
+    public void getPulseData() {
+        Query query = mRef.child("pulses").orderByChild("employee").equalTo(mRef.getAuth().getUid());
+        query.addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                pulseList.clear();
+                for(DataSnapshot child : dataSnapshot.getChildren()) {
+                    pulseList.add(child.getValue(Pulse.class));
+                }
+
+                // Here we know pulsesList is filled
+                //trigger update of weekview;
+                mWeekView.notifyDatasetChanged();
+            }
+
+            @Override
+            public void onCancelled(FirebaseError firebaseError) {}
+        });
+    }
+
+    @Override
+    public List<?extends WeekViewEvent> onMonthChange(int newYear, int newMonth) {
+        List<WeekViewEvent> events = new ArrayList<>();
+        for (Pulse pulse : pulseList) {
+            Calendar startTime = Calendar.getInstance();
+            startTime.setTimeInMillis(pulse.getCheckin());
+            int pulseMonth = startTime.get(Calendar.MONTH);
+            int pulseYear = startTime.get(Calendar.YEAR);
+            if (newMonth == pulseMonth && newYear == pulseYear) {
+                // create calendar instance of checkin
+                startTime.setTimeInMillis(pulse.getCheckin());
+                // create calendar instance of checkout
+                Calendar endTime = Calendar.getInstance();
+                Calendar now = Calendar.getInstance();
+                if (pulse.getCheckout() == 0) {
+                    endTime.setTimeInMillis(now.getTimeInMillis());
+                } else {
+                    endTime.setTimeInMillis(pulse.getCheckout());
+                }
+                // add event to the eventlist
+                WeekViewEvent event = new WeekViewEvent(1,getEventTitle(pulse),startTime,endTime);
+                event.setColor(ContextCompat.getColor(getContext(),R.color.colorPrimary));
+                events.add(event);
+            }
+        }
+        return events;
+    }
+    @Override
+    public void onEventClick(WeekViewEvent event, RectF eventRect) {
+
+    }
+    @Override
+    public void onEmptyViewLongPress(Calendar calendar) {
+
+    }
+    @Override
+    public void onEventLongPress(WeekViewEvent event,RectF eventRect) {
+
+    }
+    public void setupVariables() {
+        //set activity
+        activity = (MainActivity) getActivity();
+
+        mRef = activity.getFirebaseRef();
+        pulseList = new ArrayList<>();
+        getPulseData();
+    }
+    public void setupWeekView(WeekView weekView) {
+
+        mWeekView = weekView;
+
+        // event click listener
+        mWeekView.setOnEventClickListener(this);
+
+        // The week view has infinite scrolling horizontally. We have to provide the events of a
+        // month every time the month changes on the week view.
+        mWeekView.setMonthChangeListener(this);
+
+        // Set long press listener for events.
+        mWeekView.setEventLongPressListener(this);
+
+        // Set long press listener for empty view
+        mWeekView.setEmptyViewLongPressListener(this);
+    }
+}
+
